extract control panel cell configuration to separate file
